+++ test/container-size.js
@@ -1,7 +1,5 @@
 test( 'container size', function() {
 
-  var getSize = window.getSize;
-
   // test layout that just sets size
   var Sizer = Outlayer.create( 'sizer', {
     width: 220,
@@ -26,6 +24,15 @@ test( 'container size', function() {
 
   checkSize( 220, 120 );
 
+  // disable resizing
+  layout.options.isResizingContainer = false;
+  layout.options.width = 180;
+  layout.options.height = 230;
+  layout.layout();
+  checkSize( 220, 120 );
+  layout.options.isResizingContainer = true;
+  layout.options.width = 220;
+  layout.options.height = 120;
 
   if ( layout.size.isBorderBox ) {
     elem.style.padding = '10px 20px 30px 40px';